backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "patentlypatent-backend"
version = "0.5.0"
description = "PatentlyPatent backend — FastAPI + SQLite + SSE + Anthropic + 智慧芽"
requires-python = ">=3.10"
authors = [{ name = "ChuangLee" }]
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "sse-starlette>=2.1",
    "httpx>=0.27",
    "anthropic>=0.40",
    "python-docx>=1.1",
    "aiofiles>=24",
]

[tool.setuptools.packages.find]
include = ["app*"]
